requirements.txt
streamlit
numpy
Pillow
qrcode


